Make trash can visible only on movie list (page)
