+++ src/components/MovieList/MovieList.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components/macro'
 
-import Button from '../Button/Button'
+import MovieCard from '../MovieCard/MovieCard'
 
 import 'css.gg/icons/css/film.css'
 
@@ -12,6 +12,8 @@ export default function MovieList({
   textAlign,
   background,
   onHandleDelete,
+  display,
+  right,
 }) {
   return (
     <MovieBorder
@@ -25,90 +27,18 @@ export default function MovieList({
       {movies
         .sort((a, b) => a.movieTitle > b.movieTitle)
         .map(movie => (
-          <MovieWrapper key={movie.movieTitle}>
-            <MovieTitle>{movie.movieTitle}</MovieTitle>{' '}
-            <FoodCategory>{movie.foodCategory}</FoodCategory>
-            <DeleteButton onClick={() => onHandleDelete(movie.movieTitle)}>
-              <ButtonIcon className="gg-trash" />
-            </DeleteButton>
-          </MovieWrapper>
+          <MovieCard
+            key={movie.movieTitle}
+            movieName={movie.movieTitle}
+            movieFoodCategory={movie.foodCategory}
+            onHandleDelete={onHandleDelete}
+            display={display}
+            right={right}
+          />
         ))}
     </MovieBorder>
   )
 }
-
-const MovieWrapper = styled.section`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  background: rgba(125, 132, 178, 0.3);
-  padding: 15px;
-  margin: 5px;
-  border-radius: 8px;
-  font-family: 'Quicksand', sans-serif;
-  position: relative;
-`
-const MovieTitle = styled.span`
-  word-wrap: break-word;
-  width: 170px;
-`
-const FoodCategory = styled.span`
-  position: absolute;
-  background: rgba(125, 132, 178, 0.6);
-  color: #5b639a;
-  border-radius: 8px;
-  font-size: 0.8em;
-  padding: 5px;
-  right: 40px;
-  margin-left: 20px;
-  z-index: 1;
-`
-const DeleteButton = styled(Button)`
-  background: transparent;
-  width: 10%;
-  position: relative;
-  right: -10px;
-`
-const ButtonIcon = styled.div`
-  &.gg-trash {
-    box-sizing: border-box;
-    position: relative;
-    display: block;
-    transform: scale(var(--ggs, 1));
-    width: 10px;
-    height: 12px;
-    border: 2px solid transparent;
-    box-shadow: 0 0 0 2px, inset -2px 0 0, inset 2px 0 0;
-    border-bottom-left-radius: 1px;
-    border-bottom-right-radius: 1px;
-    margin-top: 4px;
-  }
-  &.gg-trash::after,
-  &.gg-trash::before {
-    content: '';
-    display: block;
-    box-sizing: border-box;
-    position: absolute;
-  }
-  &.gg-trash::after {
-    background: currentColor;
-    border-radius: 3px;
-    width: 16px;
-    height: 2px;
-    top: -4px;
-    left: -5px;
-  }
-  &.gg-trash::before {
-    width: 10px;
-    height: 4px;
-    border: 2px solid;
-    border-bottom: transparent;
-    border-top-left-radius: 2px;
-    border-top-right-radius: 2px;
-    top: -7px;
-    left: -2px;
-  }
-`
 
 export const MovieBorder = styled.div`
   border: 2px solid #7d84b2;
